Introduce compiled model property manager
