+++ src/plugins/intel_npu/src/plugin/include/compiled_model_property_manager.hpp
@@ -4,14 +4,17 @@
 
 #pragma once
 
+#include <map>
 #include <memory>
+#include <mutex>
 #include <optional>
 #include <string>
+#include <vector>
 
 #include "intel_npu/common/filtered_config.hpp"
 #include "intel_npu/common/igraph.hpp"
 #include "intel_npu/utils/logger/logger.hpp"
-#include "properties.hpp"
+#include "property_registration.hpp"
 
 namespace intel_npu {
 
@@ -26,16 +29,20 @@ public:
     ov::Any getProperty(const std::string& name) const;
 
     const FilteredConfig& getConfig() const {
-        return _properties.getConfig();
+        return _config;
     }
 
 private:
+    void registerProperties();
     std::string buildRuntimeRequirements() const;
 
+    mutable std::mutex _mutex;
+    FilteredConfig _config;
     std::shared_ptr<IGraph> _graph;
     std::optional<int64_t> _batchSize;
     Logger& _logger;
-    mutable Properties _properties;
+    std::map<std::string, PropertyDescriptor> _properties;
+    std::vector<ov::PropertyName> _supportedProperties;
 };
 
 }  // namespace intel_npu
